move event sending out of main
